head/ui: promote the OpenPBR editor, delete the legacy Appearances tab (M46-F07)
The Materials window had two parallel appearance tabs (metallic-roughness
"Appearances" and the full-spec "OpenPBR"), backing the now-deleted dual
model chain. Deletes the legacy tab and its editor entirely; renames the
OpenPBR editor's session interface, functions, and selection state to
drop the prefix (appearanceEditorSession, drawAppearanceTabContent,
selectedAppearance) now that it's the only appearance editor.

Lowers the head/ui session-coupling ratchet from 440 to 436: removing the
duplicate tab's own Session references is a real coupling reduction, not
a regression.
