+++ archguard/head_session_ratchet_test.go
@@ -21,7 +21,7 @@ import (
 // the compile-time interface assertions that PROVE a consumer interface, and doc comments
 // that merely mention the type). Lower it when a conversion removes references; never raise
 // it — a raise means a new widget took the whole session instead of a slim interface.
-const headSessionParamPin = 440
+const headSessionParamPin = 436
 
 func TestHeadSessionCouplingRatchet(t *testing.T) {
 	got := countSessionCoupling(t)
